Khai

+++ restaurant-be/src/main/java/com/ktpm1/restaurant/services/impls/FoodServiceImpl.java
@@ -1,19 +1,29 @@
 package com.ktpm1.restaurant.services.impls;
 
+import com.ktpm1.restaurant.dtos.events.EventHistoryFoodCreate;
 import com.ktpm1.restaurant.dtos.request.FoodRequest;
+import com.ktpm1.restaurant.dtos.response.FoodDetailAndRelated;
 import com.ktpm1.restaurant.models.FileUpload;
 import com.ktpm1.restaurant.models.Food;
+import com.ktpm1.restaurant.models.SessionTime;
+import com.ktpm1.restaurant.models.User;
 import com.ktpm1.restaurant.repositories.CategoryRepository;
 import com.ktpm1.restaurant.repositories.FoodRepository;
+import com.ktpm1.restaurant.repositories.UserRepository;
 import com.ktpm1.restaurant.services.FileUploadService;
 import com.ktpm1.restaurant.services.FoodService;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.time.LocalTime;
+import java.util.ArrayList;
+import java.util.List;
 
 @Service
 public class FoodServiceImpl implements FoodService {
@@ -23,6 +33,10 @@ public class FoodServiceImpl implements FoodService {
     private CategoryRepository categoryRepository;
     @Autowired
     private FileUploadService fileUploadService;
+    @Autowired
+    private UserRepository userRepository;
+    @Autowired
+    private KafkaTemplate<String, Object> kafkaTemplate;
 
     @Override
     public Food createFood(FoodRequest foodRequest) {
@@ -31,13 +45,28 @@ public class FoodServiceImpl implements FoodService {
         food.setDescription(foodRequest.getDescription());
         food.setPrice(foodRequest.getPrice());
         food.setCategory(categoryRepository.findById(foodRequest.getCategoryId()).orElse(null));
+        food.setSessionTime(foodRequest.getSessionTime());
 
         MultipartFile file = foodRequest.getFile();
         if (file != null && !file.isEmpty()) {
             try {
                 FileUpload fileUpload = fileUploadService.uploadFile(file);
-                food.setImage(fileUpload);
+                food.setImageCode(fileUpload.getFileCode());
             } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+        }
+
+        List<MultipartFile> files = foodRequest.getFiles();
+        if (files != null) {
+            List<String> imageList = new ArrayList<>();
+            try {
+                for (MultipartFile multipartFile : files) {
+                    FileUpload fileUpload = fileUploadService.uploadFile(multipartFile);
+                    imageList.add(fileUpload.getFileCode());
+                }
+                food.setImageList(imageList);
+            } catch (Exception e) {
                 throw new RuntimeException(e);
             }
         }
@@ -53,20 +82,47 @@ public class FoodServiceImpl implements FoodService {
                     food1.setDescription(foodRequest.getDescription());
                     food1.setPrice(foodRequest.getPrice());
                     food1.setCategory(categoryRepository.findById(foodRequest.getCategoryId()).orElse(null));
+                    food1.setSessionTime(foodRequest.getSessionTime());
 
                     MultipartFile file = foodRequest.getFile();
 
                     Food updatedFood = null;
 
                     if (file != null && !file.isEmpty()) {
-                        String fileCode = food1.getImage().getFileCode();
+                        String fileCode = null;
+                        fileCode = food1.getImageCode();
                         try {
                             FileUpload fileUpload = fileUploadService.uploadFile(file);
-                            food1.setImage(fileUpload);
-                            updatedFood = foodRepository.save(food1);
-                            fileUploadService.deleteFileByFileCode(fileCode);
+                            food1.setImageCode(fileUpload.getFileCode());
+                            if (fileCode != null) {
+                                fileUploadService.deleteFileByFileCode(fileCode);
+                            }
                         } catch (IOException e) {
                             throw new RuntimeException(e);
+                        }
+
+                        List<MultipartFile> files = foodRequest.getFiles();
+                        if (files != null) {
+                            List<String> imageListOfFood = food1.getImageList();
+                            List<String> imageList = new ArrayList<>();
+                            try {
+                                for (MultipartFile multipartFile : files) {
+                                    FileUpload fileUpload = fileUploadService.uploadFile(multipartFile);
+                                    imageList.add(fileUpload.getFileCode());
+                                }
+                                food1.setImageList(imageList);
+                                updatedFood = foodRepository.save(food1);
+                            } catch (Exception e) {
+                                throw new RuntimeException(e);
+                            }
+
+                            for(String imageCode : imageListOfFood) {
+                                try {
+                                    fileUploadService.deleteFileByFileCode(imageCode);
+                                } catch (IOException e) {
+                                    throw new RuntimeException(e);
+                                }
+                            }
                         }
                     }
                     return updatedFood;
@@ -87,5 +143,55 @@ public class FoodServiceImpl implements FoodService {
     @Override
     public Page<Food> getAllFoods(Pageable pageable, String keyword) {
         return foodRepository.getAllFood(pageable, keyword);
+    }
+
+    @Override
+    public List<Food> getAllFoods(String keyword) {
+        return foodRepository.getAllFood(keyword);
+    }
+
+    @Override
+    public List<Food> getAllFoodsBySessionTime() {
+        LocalTime now = LocalTime.now();
+
+        if (now.isAfter(LocalTime.of(6, 0)) && now.isBefore(LocalTime.of(10, 0))) {
+            return foodRepository.findAllBySessionTime(SessionTime.MORNING);
+        } else if (now.isAfter(LocalTime.of(10, 0)) && now.isBefore(LocalTime.of(14, 0))) {
+            return foodRepository.findAllBySessionTime(SessionTime.MORNING);
+        } else if (now.isAfter(LocalTime.of(14, 0)) && now.isBefore(LocalTime.of(18, 0))) {
+            return foodRepository.findAllBySessionTime(SessionTime.EVENING);
+        } else if (now.isAfter(LocalTime.of(18, 0)) && now.isBefore(LocalTime.of(22, 0))) {
+            return foodRepository.findAllBySessionTime(SessionTime.NIGHT);
+        }
+        return foodRepository.findAllBySessionTime(SessionTime.MORNING);
+    }
+
+    @Override
+    public FoodDetailAndRelated getFoodDetailAndRelated(Long id, String username) {
+        Food food = foodRepository.findById(id).orElse(null);
+        if (food == null) {
+            return null;
+        }
+
+        List<Food> relatedFoods = foodRepository.findByCategoryAndIdNot(food.getCategory(), id);
+        User user = userRepository.findByUsername(username);
+        // kafkaTemplate.send("food-view", EventHistoryFoodCreate.builder().food(food).user(user).build());
+        return FoodDetailAndRelated.builder().food(food).relatedFoods(relatedFoods).build();
+    }
+
+    @Override
+    public FoodDetailAndRelated getFoodDetailAndRelated(Long id) {
+        Food food = foodRepository.findById(id).orElse(null);
+        if (food == null) {
+            return null;
+        }
+
+        List<Food> relatedFoods = foodRepository.findByCategoryAndIdNot(food.getCategory(), id);
+        return FoodDetailAndRelated.builder().food(food).relatedFoods(relatedFoods).build();
+    }
+
+    @Override
+    public Food getFoodById(Long id) {
+        return foodRepository.findById(id).orElse(null);
     }
 }
